requirements.txt
dash
dash_bootstrap_components
dash_svg
pandas
flask_caching
redis
gunicorn
psycopg2-binary
python-dotenv